(chore) Norminette fix
