+++ src/executor_utils.c
@@ -1,0 +1,77 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   executor_utils.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: Juliany Bernardo <julberna@student.42sp    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/02/23 17:09:16 by Juliany Ber       #+#    #+#             */
+/*   Updated: 2024/02/23 17:12:19 by Juliany Ber      ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "./includes/minishell.h"
+
+void	free_array_of_strings(char **array, int size)
+{
+	int	index;
+
+	if (array == NULL)
+		return ;
+	index = 0;
+	while (index <= size)
+	{
+		free(array[index]);
+		index++;
+	}
+}
+
+char	**stringfy(t_word *wl)
+{
+	char	**array;
+	int		size;
+	int		index;
+
+	size = get_size(wl) + 1;
+	array = ft_calloc(size, sizeof(char *));
+	index = 0;
+	while (wl->next != NULL)
+	{
+		array[index] = ft_strdup(wl->word);
+		index++;
+		wl = wl->next;
+	}
+	array[index] = ft_strdup(wl->word);
+	return (array);
+}
+
+int	get_size(t_word *wl)
+{
+	int	nelem;
+
+	nelem = 1;
+	while (wl->next != NULL)
+	{
+		nelem++;
+		wl = wl->next;
+	}
+	return (nelem);
+}
+
+void	close_all_fds(void)
+{
+	int	i;
+
+	i = 0;
+	while (i < 1024)
+	{
+		close(i);
+		i++;
+	}
+}
+
+void	handle_null_pathname(t_meta *meta)
+{
+	ft_putstr_fd("minishell: command not found\n", STDERR_FILENO);
+	add_or_upd_ht_entry("?", "127", meta->hash);
+}
